Now supports files with no audio

+++ wackywebm.js
@@ -30,7 +30,6 @@ const input = `"${process.argv.slice(2).join(" ")}"`;
 const delta = 2;
 
 async function main() {
-	if(!input) return console.log("WackyWebM, by OIRNOIR#0032\nUsage: node wackywebm <input_file>");
 	console.log(`Input file: ${input}\nUsing minimum w/h ${delta}px\nGetting framerate...`);
 	const inputFramerateRes = await execa(`ffprobe -v error -select_streams v -of default=noprint_wrappers=1:nokey=1 -show_entries stream=r_frame_rate ${input}`);
 	const inputFramerate = inputFramerateRes.stdout.replace(/\n/, "");
@@ -41,9 +40,15 @@ async function main() {
 	console.log(`Got width of ${maxWidth}! Getting height...`);
 	const maxHeightRes = await execa(`ffprobe -v error -select_streams v -of default=noprint_wrappers=1:nokey=1 -show_entries stream=height ${input}`);
 	const maxHeight = Number(maxHeightRes.stdout.replace(/\n/, ""));
-	console.log(`Got height of ${maxHeight}! Splitting audio to temporary file...`);
+	console.log(`Got height of ${maxHeight}! Checking if there is audio...`);
+	const hasAudioRes = await execa(`ffprobe -loglevel error -show_entries stream=codec_type -of default=noprint_wrappers=1:nokey=1 ${input}`);
+	const hasAudio = hasAudioRes.stdout.includes("audio");
+	console.log(`This file ${hasAudio ? "has" : "does not have"} audio!`);
 	let lines = [];
-	await execa(`ffmpeg -y -i ${input} -vn -c:a libvorbis tempAudio.webm`);
+	if (hasAudio) {
+		console.log("Splitting audio into temporary file...");
+		await execa(`ffmpeg -y -i ${input} -vn -c:a libvorbis tempAudio.webm`);
+	}
 	console.log(`Creating temporary directories...`);
 	try {
 		await execa(`mkdir tempFrames;mkdir tempResizedFrames`);
@@ -57,7 +62,7 @@ async function main() {
 	for (const file of fs.readdirSync(`${__dirname}/tempFrames/`).filter(f => f.endsWith("png")).sort((a, b) => Number(a.split("/").slice(-1)[0].split(".")[0]) - Number(b.split("/").slice(-1)[0].split(".")[0]))) {
 		const width = index == 0 ? maxWidth : (Math.floor(Math.random() * (maxWidth - delta)) + delta);
 		const height = index == 0 ? maxHeight : (Math.floor(Math.random() * (maxHeight - delta)) + delta);
-		const bouncespersecond = 1.9;
+		const bouncespersecond = 2;
 		//const width = maxWidth;
 		//const height = index == 0 ? maxHeight : (Math.floor(Math.abs(Math.cos(index / (decimalInputFramerate / bouncespersecond) * Math.PI) * (maxHeight - delta))) + delta);
 		await execa(`ffmpeg -y -i ${__dirname}/tempFrames/${file} -c:v vp8 -b:v 1M -crf 10 -vf scale=${width}x${height} -aspect ${width}:${height} -r ${inputFramerate} -f webm ${__dirname}/tempResizedFrames/${file}.webm`);
@@ -71,22 +76,26 @@ async function main() {
 	console.log("Writing concat file...");
 	await writeFileAsync("./tempConcatList.txt", lines.join("\n"));
 	console.log("Combining webm files into a single webm...");
-	await execa(`ffmpeg -y -f concat -safe 0 -i tempConcatList.txt -c copy tempVideo.webm`);
-	console.log("Creating final webm...");
-	await execa(`ffmpeg -y -i tempVideo.webm -i tempAudio.webm -c copy out.webm`);
-	console.log("Done!");
-		console.log("Removing temporary files...");
-		for (const file of fs.readdirSync(`${__dirname}/tempFrames/`)) {
-			fs.unlinkSync(`${__dirname}/tempFrames/${file}`);
-		}
-		fs.rmdirSync(__dirname + "/tempFrames");
-		for (const file of fs.readdirSync(`${__dirname}/tempResizedFrames/`)) {
-			fs.unlinkSync(`${__dirname}/tempResizedFrames/${file}`);
-		}
-		fs.rmdirSync(__dirname + "/tempResizedFrames");
+	await execa(`ffmpeg -y -f concat -safe 0 -i tempConcatList.txt -c copy ${hasAudio ? "tempVideo.webm" : "out.webm"}`);
+	if (hasAudio) {
+		console.log("Merging video and audio...");
+		await execa(`ffmpeg -y -i tempVideo.webm -i tempAudio.webm -c copy out.webm`);
+	}
+	console.log("Finished making the file! Cleaning up...");
+	for (const file of fs.readdirSync(`${__dirname}/tempFrames/`)) {
+		fs.unlinkSync(`${__dirname}/tempFrames/${file}`);
+	}
+	fs.rmdirSync(__dirname + "/tempFrames");
+	for (const file of fs.readdirSync(`${__dirname}/tempResizedFrames/`)) {
+		fs.unlinkSync(`${__dirname}/tempResizedFrames/${file}`);
+	}
+	fs.rmdirSync(__dirname + "/tempResizedFrames");
+	fs.unlinkSync(__dirname + "/tempConcatList.txt");
+	if (hasAudio) {
 		fs.unlinkSync(__dirname + "/tempAudio.webm");
 		fs.unlinkSync(__dirname + "/tempVideo.webm");
-		fs.unlinkSync(__dirname + "/tempConcatList.txt");
+	}
+	console.log("Done!");
 }
 
 void main();
